Init commit with basic web app
This commit includes the chef setup logic to get the python environment
correctly setup, as well as the initial commits for the web application
server.
The server is serving a landing page on the root route. The landing
page has a drop down and textarea for inputs. The submit button triggers
an async POST to the server that creates the new comment in the
applications database.