Get application running as service with https
This commit includes the changes necessary to have the application
running as a systemd managed service behind uwsgi. This change also
includes the basics of getting https running on the application with
nginx terminating the tls connection and forwarding the connection onto
the uwsgi application. Initial certbot steps required for TLS
certificates are also included. Both nginx and certbot setups will get
expanded on.