Add Asana key programmatically
Preferential usage would be to get the Asana API key out of a secrets
store such as Azure Key Vault. But that is not possible to do spanning
multiple Azure accounts with any users wanting to use this repo to spin
up a web app like this. Therefore we use terraform vars as input, and
cloud init template vars to pass in the Asana key.