Create new Server script
This change creates a shell script that can be run to create a brand new
server that would have this solution running on it. The script requires
you to pass the Asana key in as a parameter (can be forced) and that key
gets installed on the freshly deployed server. This script was run from
within an Azure cloud shell successfully on many occassions. The user
running this script needs to ensure that they are logged into the Azure
cli tool (aka run `az login`)