Get Asana task using file key and cron trigger
The script that creates the task on Asana with the summary of the last
day's worth of feedback is now using the on disk Asana key that gets
created at run time (preference would be to use a secret store rather).
The script is also added to cron to get triggered automatically once a
day.